+++ src/input_processing/single_axis.rs
@@ -132,6 +132,22 @@ crate::define_dynamic_input_processing_pipeline!(
     processor_type: AxisProcessor
 );
 
+/// A trait for processors that can be merged with other [`AxisProcessor`]s.
+pub trait MergeAxisProcessor {
+    /// Merges this processor with another [`AxisProcessor`].
+    fn merge_processor(self, processor: impl AxisProcessor) -> Self;
+}
+
+impl MergeAxisProcessor for Box<dyn AxisProcessor> {
+    fn merge_processor(self, processor: impl AxisProcessor) -> Self {
+        let pipeline = match Reflect::as_any(&*self).downcast_ref::<AxisProcessingPipeline>() {
+            Some(pipeline) => pipeline.clone(),
+            None => AxisProcessingPipeline(vec![self]),
+        };
+        Box::new(pipeline.with(processor))
+    }
+}
+
 // endregion pipeline
 
 // region inversion
@@ -740,6 +756,27 @@ mod tests {
         let pipeline = InvertedThenDouble;
         assert_eq!(pipeline.process(2.0), -4.0);
         assert_eq!(pipeline.process(-1.0), 2.0);
+    }
+
+    #[test]
+    fn test_merge_axis_processor() {
+        let first = AxisSensitivity(2.0);
+        let first_boxed: Box<dyn AxisProcessor> = Box::new(first);
+
+        let second = AxisSensitivity(3.0);
+        let merged_second = first_boxed.merge_processor(second);
+        let expected = AxisProcessingPipeline::default().with(first).with(second);
+        let expected_boxed: Box<dyn AxisProcessor> = Box::new(expected);
+        assert_eq!(merged_second, expected_boxed);
+
+        let third = AxisSensitivity(4.0);
+        let merged_third = merged_second.merge_processor(third);
+        let expected = AxisProcessingPipeline::default()
+            .with(first)
+            .with(second)
+            .with(third);
+        let expected_boxed: Box<dyn AxisProcessor> = Box::new(expected);
+        assert_eq!(merged_third, expected_boxed);
     }
 
     #[test]